Updated saved event controller, layout of buttons in main handlebars and removed auth from search.
Updated the savedEventController.js to directly call the discovery API.  Save event button saves event information through the models.  The unique ID is grabbed from the models and makes a get call then a post with the event array specifically to avoid loss of data during model transfer.

Removed auth requirement to search when logged in only, doesn't make sense that a user would need to be logged in to search events generally.

Changed the layouts of the buttons.